# backend/cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple backend server for authenticated web services",
        "title": "Authenticated Web Services",
        "termsOfService": "http://localhost:8000",
        "contact": {
            "name": "API Support",
            "email": "meichang@arch.nctu.edu.tw"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "this is a method to show the server alive",
                "summary": "show status of server",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.HelloPayload"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "authenticate user for login the server and get JWT tokens with openid-connect by keycloak",
                "summary": "authenticate user for login",
                "parameters": [
                    {
                        "description": "user name registered at keycloak",
                        "name": "Username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.LoginPayload"
                        }
                    },
                    {
                        "description": "Password registered at keycloak",
                        "name": "Password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.LoginPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.TokenPair"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "description": "logout user",
                "summary": "logout user",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/swissdata/datalist/{organization_name}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get opendata.swiss's data list with JWT token by organization name",
                "summary": "get opendata.swiss's data list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "organization name",
                        "name": "organization_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.SwissData"
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/swissdata/organizations": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get opendata.swiss's organization list with JWT token",
                "summary": "get opendata.swiss's organization list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Organization"
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.HelloPayload": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "main.LoginPayload": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "main.Organization": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "main.SwissData": {
            "type": "object",
            "properties": {
                "download_url": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "issued": {
                    "type": "string"
                },
                "maintainer": {
                    "type": "string"
                },
                "maintainer_email": {
                    "type": "string"
                },
                "owner_org": {
                    "type": "string"
                }
            }
        },
        "main.TokenPair": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer"
                },
                "refresh_expires_in": {
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}